iot_api/requirements.txt
fastapi==0.110.2
passlib==1.7.4
pycryptodome==3.20.0
pydantic==2.7.1
pymongo==4.6.3
python_jose==3.3.0
pytz==2024.1
uvicorn==0.29.0
python-multipart